读文件/requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
chardet==2.3.0
Django==1.11.3
docx2txt==0.6
EbookLib==0.15
freeze==1.0.10
lxml==4.3.4
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
pocketsphinx==0.1.3
python-dateutil==2.8.0
python-docx==0.8.10
python-pptx==0.6.5
pytz==2019.1
six==1.10.0
SpeechRecognition==3.6.3
xlrd==1.0.0
XlsxWriter==1.1.8
